alksjd

+++ point.h
@@ -18,11 +18,11 @@ public:
         x = 0; y = 0; z = 0; seg1 = 0; seg2 = 0;
     }
     
-    Point(float my_x, float my_y){
+    Point(double my_x, double my_y){
         x = my_x; y = my_y; z = 0; seg1 = 0; seg2 = 0;
     }
     
-    Point(int my_x, int my_y, int my_z){
+    Point(double my_x, double my_y, double my_z){
         x = my_x; y = my_y; z = my_z; seg1 = 0; seg2 = 0;
     }
     
@@ -30,28 +30,28 @@ public:
         return (x < p->getX() || (x == p->getX() && y > p->getY()));
     }
     
-    void setX(float x_){
+    void setX(double x_){
         x = x_;
     }
     
-    void setY(float y_){
+    void setY(double y_){
         y = y_;
     }
     
-    void setZ(float z_){
+    void setZ(double z_){
         z = z_;
     }
     
-    float getX() {return x;}
-    float getY() {return y;}
-    float getZ() {return z;}
+    double getX() {return x;}
+    double getY() {return y;}
+    double getZ() {return z;}
     Segment* getSeg1() {return seg1;}
     Segment* getSeg2() {return seg2;}
     void setSeg1(Segment* s) {seg1 = s;}
     void setSeg2(Segment* s) {seg2 = s;}
     
 private:
-    float x, y, z; //the points coordinates
+    double x, y, z; //the points coordinates
     Segment* seg1; //a segment that has this point as one of its endpoints
     Segment* seg2; //a segment that has this point as one of its endpoints
